requirements.txt
streamlit>=1.30.0
boto3>=1.34.0
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.1.0
langchain>=0.3.0
langchain-aws>=0.2.0
langgraph>=0.2.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
